# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: ERP Backend API (Phase 1 + Phase 2 Inc1 + Inc2 LV §9 + FIN-0 Finance Stubs)
  version: 1.7.0-password-reset
  description: |
    **PWA / Browser-Anbindung**
    - **CORS**: Umgebungsvariable `CORS_ORIGINS` — kommagetrennte, exakte Origins (z. B. `https://app.example.com,http://localhost:5173`). Leer oder unset: keine `Access-Control-*`-Header (gleiche Origin oder Server-zu-Server).
    - **Auth**: `Authorization: Bearer <signed-v1-token>` (siehe `bearerToken`); optionaler Header `X-Tenant-Id` muss zum Tenant im Token passen, sonst `TENANT_SCOPE_VIOLATION`. **Passwort-Login:** `POST /auth/login` mit JSON `{ tenantId, email, password }` — siehe `docs/authentication-login.md`. **Passwort vergessen (öffentlich):** `POST /auth/request-password-reset`, `POST /auth/confirm-password-reset` — nur mit Postgres; Link `#/password-reset?token=…` (Basis `ERP_PUBLIC_APP_BASE_URL`); Produktion: SMTP + Basis-URL Pflicht. **Benutzerverwaltung:** `GET|POST /users`, `PATCH /users/{userId}` inkl. optional **E-Mail** (nur Rolle **ADMIN**, nur mit Postgres) — Audit-Ereignisse `USER` / `USER_CREATED` / `USER_UPDATED`. **Betrieb:** `AUTH_TOKEN_SECRET` (min. 32 Zeichen bei `NODE_ENV=production`); ohne Secret kein Start — siehe Repo-`.env.example`. Nur lokaler Demo-Modus: `ERP_ALLOW_INSECURE_DEV_AUTH=1` und nicht `production` (niemals deployen).
...
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: signed-v1-token
  parameters:
    TenantScopeHeader:
      in: header
      name: x-tenant-id
      required: false
      schema:
        type: string
        format: uuid
      description: Optionaler Tenant-Scope, muss mit Token-Tenant identisch sein.
...
  schemas:
...
    TransitionLvVersionRequest:
      type: object
      required: [nextStatus, reason]
      properties:
        nextStatus:
          type: string
          enum: [FREIGEGEBEN, ARCHIVIERT]
        reason:
          type: string
          minLength: 5
...
paths:
...
  /lv/catalogs:
    post:
      description: Neuer LV-Katalog mit erster Version (ENTWURF) und Minimal-Skelett §9.
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLvCatalogRequest'
      responses:
        '201':
          description: lvCatalogId, lvVersionId, samplePositionId
        '403':
          description: AUTH_ROLE_FORBIDDEN
        '401':
          description: UNAUTHORIZED
...
  /lv/versions/{lvVersionId}/status:
    post:
      description: LV-Version Status ENTWURF→FREIGEGEBEN→ARCHIVIERT (nur aktuelle Version).
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: path
          name: lvVersionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionLvVersionRequest'
      responses:
        '200':
          description: LvVersion
        '409':
          description: LV_STATUS_TRANSITION_FORBIDDEN | LV_VERSION_NOT_CURRENT | LV_STRUCTURE_LOCKED
        '403':
          description: AUTH_ROLE_FORBIDDEN
  /lv/versions/{lvVersionId}/nodes:
    post:
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: path
          name: lvVersionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddLvStructureNodeRequest'
      responses:
        '201':
          description: LvStructureNode
        '409':
          description: LV_STRUCTURE_LOCKED | LV_HIERARCHY_INVALID
        '404':
          description: LV_NODE_NOT_FOUND
  /lv/nodes/{nodeId}/editing-text:
    patch:
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: path
          name: nodeId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLvNodeEditingRequest'
      responses:
        '200':
          description: LvStructureNode
        '409':
          description: LV_STRUCTURE_LOCKED
  /lv/versions/{lvVersionId}/positions:
    post:
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: path
          name: lvVersionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddLvPositionRequest'
      responses:
        '201':
          description: LvPosition
        '409':
          description: LV_STRUCTURE_LOCKED | LV_HIERARCHY_INVALID
  /lv/positions/{positionId}:
    patch:
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: path
          name: positionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchLvPositionRequest'
      responses:
        '200':
          description: LvPosition
        '404':
          description: LV_POSITION_NOT_FOUND
        '409':
          description: LV_STRUCTURE_LOCKED
        '403':
          description: AUTH_ROLE_FORBIDDEN
  /measurements:
    post:
      description: |
        Neues Aufmass mit erster Version (ENTWURF). AuthZ SoT-Aktion MEASUREMENT_CREATE.
        Siehe GET /documents/{id}/allowed-actions?entityType=MEASUREMENT_VERSION für Folgeaktionen auf der aktuellen Version.
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMeasurementRequest'
      responses:
        '201':
          description: measurementId und measurementVersionId
        '403':
          description: AUTH_ROLE_FORBIDDEN
        '422':
          description: MEASUREMENT_POSITIONS_REQUIRED / Zod
        '401':
          description: Authentifizierung fehlgeschlagen
  /measurements/status:
    post:
      description: Statusübergang nur auf der aktuellen Aufmassversion (§5.4). Kritische Übergänge mit AuditEvent.
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionMeasurementStatusRequest'
      responses:
        '200':
          description: Aktualisierte MeasurementVersion
        '409':
          description: MEASUREMENT_STATUS_TRANSITION_FORBIDDEN | MEASUREMENT_VERSION_NOT_CURRENT
        '403':
          description: AUTH_ROLE_FORBIDDEN
        '404':
          description: MEASUREMENT_NOT_FOUND
        '401':
          description: Authentifizierung fehlgeschlagen
...
  /measurements/{measurementVersionId}/positions:
    post:
      description: Ersetzt Positionen nur bei aktueller Version im Status ENTWURF (§5.4).
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: path
          name: measurementVersionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeasurementPositionsRequest'
      responses:
        '200':
          description: positions Array
        '409':
          description: MEASUREMENT_POSITION_EDIT_FORBIDDEN | MEASUREMENT_VERSION_NOT_CURRENT
        '403':
          description: AUTH_ROLE_FORBIDDEN
        '404':
          description: MEASUREMENT_NOT_FOUND
        '401':
          description: Authentifizierung fehlgeschlagen
  /measurements/{measurementVersionId}:
    get:
      description: Liest Aufmassversion inkl. Positionen (tenant-gefiltert).
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: path
          name: measurementVersionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: version, measurementId, positions
        '404':
          description: MEASUREMENT_NOT_FOUND
        '401':
          description: Authentifizierung fehlgeschlagen
...
  /offers/{offerId}/supplements:
    post:
      description: |
        Aktion OFFER_CREATE_SUPPLEMENT (Minimal-API). AuthZ identisch zu GET /documents/{id}/allowed-actions
        für dieselbe Rolle (OFFER_CREATE_SUPPLEMENT in allowedActions).
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: path
          name: offerId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
      responses:
        '201':
          description: Nachtragsangebot-Version (ENTWURF) erstellt
        '409':
          description: SUPPLEMENT_BASE_NOT_ACCEPTED (Basisversion nicht ANGENOMMEN)
        '403':
          description: AUTH_ROLE_FORBIDDEN oder TENANT_SCOPE_VIOLATION
        '404':
          description: OFFER_NOT_FOUND oder OFFER_VERSION_NOT_FOUND
        '401':
          description: Authentifizierung fehlgeschlagen
  /supplements/status:
    post:
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
      responses:
        '200':
          description: Nachtragsstatus geaendert (IN_FREIGABE|FREIGEGEBEN|VERSENDET|BEAUFTRAGT|ABGELEHNT|ARCHIVIERT)
        '409':
          description: SUPPLEMENT_STATUS_TRANSITION_FORBIDDEN
        '404':
          description: SUPPLEMENT_NOT_FOUND
        '403':
          description: AUTH_ROLE_FORBIDDEN oder TENANT_SCOPE_VIOLATION
        '401':
          description: Authentifizierung fehlgeschlagen
  /supplements/{supplementVersionId}/billing-impact:
    post:
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: path
          name: supplementVersionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Abrechnungswirkung mit Nachtragsreferenz angewendet
        '409':
          description: SUPPLEMENT_BILLING_EFFECT_FORBIDDEN
        '404':
          description: SUPPLEMENT_NOT_FOUND oder EXPORT_ENTITY_NOT_FOUND
        '422':
          description: TRACEABILITY_LINK_MISSING
        '403':
          description: AUTH_ROLE_FORBIDDEN oder TENANT_SCOPE_VIOLATION
        '401':
          description: Authentifizierung fehlgeschlagen
  /supplements/{supplementVersionId}:
    get:
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: path
          name: supplementVersionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Nachtragsversion inklusive unveraenderlicher Basisreferenz
        '404':
          description: SUPPLEMENT_NOT_FOUND
        '403':
          description: TENANT_SCOPE_VIOLATION
        '401':
          description: Authentifizierung fehlgeschlagen
  /offers/status:
    post:
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
      responses:
        '200':
          description: Status geaendert (IN_FREIGABE|FREIGEGEBEN|VERSENDET|ANGENOMMEN|ABGELEHNT)
        '409':
          description: STATUS_TRANSITION_FORBIDDEN
        '403':
          description: AUTH_ROLE_FORBIDDEN oder TENANT_SCOPE_VIOLATION
        '401':
          description: Authentifizierung fehlgeschlagen
  /exports:
    post:
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '201':
          description: Exportlauf erfolgreich
        '422':
          description: EXPORT_PREFLIGHT_FAILED oder TRACEABILITY_LINK_MISSING oder TRACEABILITY_FIELD_MISMATCH
        '403':
          description: AUTH_ROLE_FORBIDDEN oder TENANT_SCOPE_VIOLATION
        '401':
          description: Authentifizierung fehlgeschlagen
  /documents/{id}/allowed-actions:
    get:
      description: |
        Liefert erlaubte actionIds für das Dokument (Backend-SoT). Kein separates entityType=OFFER in Phase 1.
        Beispiel nach Annahme: bei OFFER_VERSION im Status ANGENOMMEN erscheint für ADMIN/VERTRIEB/GESCHAEFTSFUEHRUNG
        OFFER_CREATE_SUPPLEMENT (nicht OFFER_CREATE_VERSION).
        Für §9 LV: `id` ist jeweils die Entitäts-ID passend zu `entityType` (LV_VERSION → lvVersionId, LV_STRUCTURE_NODE → nodeId, LV_POSITION → positionId); siehe ADR-0005.
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: entityType
          required: true
          schema:
            type: string
            enum:
              - OFFER_VERSION
              - SUPPLEMENT_VERSION
              - MEASUREMENT_VERSION
              - LV_VERSION
              - LV_STRUCTURE_NODE
              - LV_POSITION
              - INVOICE
      responses:
        '200':
          description: Erlaubte Aktionen pro Dokument als Backend-Source-of-Truth
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllowedActionsResponse'
              examples:
                offerVersionAcceptedSales:
                  summary: OFFER_VERSION ANGENOMMEN, Rolle VERTRIEB
                  value:
                    documentId: "33333333-3333-4333-8333-333333333333"
                    entityType: OFFER_VERSION
                    allowedActions: ["OFFER_CREATE_SUPPLEMENT"]
        '403':
          description: TENANT_SCOPE_VIOLATION
        '404':
          description: DOCUMENT_NOT_FOUND
        '401':
          description: Authentifizierung fehlgeschlagen
  /audit-events:
    get:
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: DSGVO-minimierte Auditansicht
        '403':
          description: FORBIDDEN_AUDIT_READ
        '401':
          description: Authentifizierung fehlgeschlagen
...